move IntoQueryFilterMap and associated plumbing into query module
