+++ domain/src/agreement.rs
@@ -1,6 +1,6 @@
 use crate::agreements::Model;
 use crate::error::Error;
-use entity_api::IntoQueryFilterMap;
+use entity_api::query::IntoQueryFilterMap;
 use entity_api::{agreements, query};
 use sea_orm::DatabaseConnection;
