+++ domain/src/coaching_relationship.rs
@@ -1,6 +1,6 @@
 use crate::coaching_relationships::Model;
 use crate::error::Error;
-use entity_api::IntoQueryFilterMap;
+use entity_api::query::IntoQueryFilterMap;
 use entity_api::{coaching_relationships, query};
 use sea_orm::DatabaseConnection;
