+++ entity_api/src/lib.rs
@@ -1,7 +1,6 @@
 use chrono::{Days, Utc};
 use password_auth::generate_hash;
-use sea_orm::{ActiveModelTrait, DatabaseConnection, Set, Value};
-use std::collections::HashMap;
+use sea_orm::{ActiveModelTrait, DatabaseConnection, Set};
 
 pub use entity::{
     actions, agreements, coachees, coaches, coaching_relationships, coaching_sessions, jwts, notes,
@@ -13,6 +12,7 @@ pub mod agreement;
 pub mod coaching_relationship;
 pub mod coaching_session;
 pub mod error;
+pub mod mutate;
 pub mod note;
 pub mod organization;
 pub mod overarching_goal;
@@ -24,87 +24,6 @@ pub(crate) fn uuid_parse_str(uuid_str: &str) -> Result<Id, error::Error> {
         source: None,
         error_kind: error::EntityApiErrorKind::InvalidQueryTerm,
     })
-}
-
-/// `QueryFilterMap` is a data structure that serves as a bridge for translating filter parameters
-/// between different layers of the application. It is essentially a wrapper around a `HashMap`
-/// where the keys are filter parameter names (as `String`) and the values are optional `Value` types
-/// from `sea_orm`.
-///
-/// This structure is particularly useful in scenarios where you need to pass filter parameters
-/// from a web request down to the database query layer in a type-safe and organized manner.
-///
-/// # Example
-///
-/// ```
-/// use sea_orm::Value;
-/// use entity_api::QueryFilterMap;
-///
-/// let mut query_filter_map = QueryFilterMap::new();
-/// query_filter_map.insert("coaching_session_id".to_string(), Some(Value::String(Some(Box::new("a_coaching_session_id".to_string())))));
-/// let filter_value = query_filter_map.get("coaching_session_id");
-/// ```
-pub struct QueryFilterMap {
-    map: HashMap<String, Option<Value>>,
-}
-
-impl QueryFilterMap {
-    pub fn new() -> Self {
-        Self {
-            map: HashMap::new(),
-        }
-    }
-
-    pub fn get(&self, key: &str) -> Option<Value> {
-        // HashMap.get returns an Option and so we need to "flatten" this to a single Option
-        self.map
-            .get(key)
-            .and_then(|inner_option| inner_option.clone())
-    }
-
-    pub fn insert(&mut self, key: String, value: Option<Value>) {
-        self.map.insert(key, value);
-    }
-}
-
-impl Default for QueryFilterMap {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// `IntoQueryFilterMap` is a trait that provides a method for converting a struct into a `QueryFilterMap`.
-/// This is particularly useful for translating data between different layers of the application,
-/// such as from web request parameters to database query filters.
-///
-/// Implementing this trait for a struct allows you to define how the fields of the struct should be
-/// mapped to the keys and values of the `QueryFilterMap`. This ensures that the data is passed
-/// in a type-safe and organized manner.
-///
-/// # Example
-///
-/// ```
-/// use entity_api::QueryFilterMap;
-/// use entity_api::IntoQueryFilterMap;
-///
-/// #[derive(Debug)]
-/// struct MyParams {
-///     coaching_session_id: String,
-/// }
-///
-/// impl IntoQueryFilterMap for MyParams {
-///     fn into_query_filter_map(self) -> QueryFilterMap {
-///         let mut query_filter_map = QueryFilterMap::new();
-///         query_filter_map.insert(
-///             "coaching_session_id".to_string(),
-///             Some(sea_orm::Value::String(Some(Box::new(self.coaching_session_id)))),
-///         );
-///         query_filter_map
-///     }
-/// }
-/// ```
-pub trait IntoQueryFilterMap {
-    fn into_query_filter_map(self) -> QueryFilterMap;
 }
 
 pub async fn seed_database(db: &DatabaseConnection) {
